requirements.txt
fastapi==0.111.0
numpy==1.26.4
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-dotenv==0.5.2
uvicorn==0.30.1
openai==1.37.1